+++ tests/test_transforms/test_decomposition/test_stl_transform.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pandas as pd
 import pytest
+from statsmodels.tsa.exponential_smoothing.ets import ETSModel
 
 from etna.datasets.tsdataset import TSDataset
 from etna.models import NaiveModel
@@ -103,7 +104,15 @@ def ts_trend_seasonal_nan_tails() -> TSDataset:
     return TSDataset(df, freq="D")
 
 
-@pytest.mark.parametrize("model", ["arima", "holt"])
+@pytest.fixture
+def not_ts_model():
+    class NotTimeSeriesModel:
+        pass
+
+    return NotTimeSeriesModel
+
+
+@pytest.mark.parametrize("model", ["arima", "holt", ETSModel])
 @pytest.mark.parametrize(
     "df_name",
     [
@@ -122,7 +131,7 @@ def test_transform_one_segment(df_name, model, request):
     np.testing.assert_allclose(df_transformed["target"], df_expected["target"], atol=0.3)
 
 
-@pytest.mark.parametrize("model", ["arima", "holt"])
+@pytest.mark.parametrize("model", ["arima", "holt", ETSModel])
 @pytest.mark.parametrize(
     "ts_name", ["ts_trend_seasonal", "ts_trend_seasonal_starting_with_nans", "ts_trend_seasonal_int_timestamp"]
 )
@@ -137,7 +146,7 @@ def test_transform_multi_segments(ts_name, model, request):
     np.testing.assert_allclose(df_transformed["target"], df_expected["target"], atol=0.3)
 
 
-@pytest.mark.parametrize("model", ["arima", "holt"])
+@pytest.mark.parametrize("model", ["arima", "holt", ETSModel])
 @pytest.mark.parametrize(
     "df_name",
     [
@@ -155,7 +164,7 @@ def test_inverse_transform_one_segment(df_name, model, request):
     pd.testing.assert_series_equal(df_inverse_transformed["target"], df["target"])
 
 
-@pytest.mark.parametrize("model", ["arima", "holt"])
+@pytest.mark.parametrize("model", ["arima", "holt", ETSModel])
 @pytest.mark.parametrize(
     "ts_name", ["ts_trend_seasonal", "ts_trend_seasonal_starting_with_nans", "ts_trend_seasonal_int_timestamp"]
 )
@@ -170,7 +179,7 @@ def test_inverse_transform_multi_segments(ts_name, model, request):
     pd.testing.assert_series_equal(df_inverse_transformed["target"], df["target"])
 
 
-@pytest.mark.parametrize("model_stl", ["arima", "holt"])
+@pytest.mark.parametrize("model_stl", ["arima", "holt", ETSModel])
 def test_forecast(ts_trend_seasonal, model_stl):
     """Test that transform works correctly in forecast."""
     transform = STLTransform(in_column="target", period=7, model=model_stl)
@@ -187,7 +196,7 @@ def test_forecast(ts_trend_seasonal, model_stl):
     ts_forecast = model.forecast(ts_future, prediction_size=3)
     ts_forecast.inverse_transform([transform])
     for segment in ts_forecast.segments:
-        np.testing.assert_allclose(ts_forecast[:, segment, "target"], ts_test[:, segment, "target"], atol=0.1)
+        np.testing.assert_allclose(ts_forecast[:, segment, "target"], ts_test[:, segment, "target"], atol=0.2)
 
 
 def test_transform_raise_error_if_not_fitted(df_trend_seasonal_one_segment):
@@ -204,7 +213,7 @@ def test_inverse_transform_raise_error_if_not_fitted(df_trend_seasonal_one_segme
         _ = transform.inverse_transform(df=df_trend_seasonal_one_segment)
 
 
-@pytest.mark.parametrize("model_stl", ["arima", "holt"])
+@pytest.mark.parametrize("model_stl", ["arima", "holt", ETSModel])
 def test_fit_transform_with_nans_in_tails(ts_trend_seasonal_nan_tails, model_stl):
     transform = STLTransform(in_column="target", period=7, model=model_stl)
     transform.fit_transform(ts=ts_trend_seasonal_nan_tails)
@@ -222,6 +231,7 @@ def test_fit_transform_with_nans_in_middle_raise_error(ts_with_nans):
     [
         STLTransform(in_column="target", period=7, model="arima"),
         STLTransform(in_column="target", period=7, model="holt"),
+        STLTransform(in_column="target", period=7, model=ETSModel),
     ],
 )
 def test_save_load(transform, ts_trend_seasonal):
@@ -233,3 +243,9 @@ def test_params_to_tune(ts_trend_seasonal):
     transform = STLTransform(in_column="target", period=7)
     assert len(transform.params_to_tune()) > 0
     assert_sampling_is_valid(transform=transform, ts=ts)
+
+
+@pytest.mark.parametrize("model_stl", [10, not_ts_model])
+def test_custom_model(model_stl):
+    with pytest.raises(ValueError, match="Model should be a string or TimeSeriesModel"):
+        transform = STLTransform(in_column="target", period=7, model=model_stl)